pdfOCR/requirements.txt
flask
pytesseract
pymupdf
opencv-python
Pillow
